requirements.txt
torch==2.1.0
dgl==2.1.0
torchdata==0.7.0
numpy==1.23.1
networkx==2.8
rdkit-pypi
scikit-learn
PyYAML==6.0
jarvis-tools==2023.12.12
pykan
